+++ common/src/main/java/com/ericsson/bss/cassandra/ecaudit/common/record/AuditRecord.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2019 Telefonaktiebolaget LM Ericsson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.ericsson.bss.cassandra.ecaudit.common.record;
+
+import java.net.InetAddress;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface AuditRecord
+{
+    Long getTimestamp();
+
+    InetAddress getClientAddress();
+
+    String getUser();
+
+    Optional<UUID> getBatchId();
+
+    Status getStatus();
+
+    AuditOperation getOperation();
+}