+++ ecaudit/src/test/java/com/ericsson/bss/cassandra/ecaudit/TestAuditAdapterFactory.java
@@ -15,16 +15,23 @@
  */
 package com.ericsson.bss.cassandra.ecaudit;
 
+import java.io.File;
 import java.lang.reflect.Field;
 import java.net.URL;
+import java.util.Collections;
+import java.util.Map;
 
+import com.google.common.collect.ImmutableMap;
+import com.google.common.io.Files;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Ignore;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
+import com.ericsson.bss.cassandra.ecaudit.config.AuditConfig;
 import com.ericsson.bss.cassandra.ecaudit.config.AuditYamlConfigurationLoader;
 import com.ericsson.bss.cassandra.ecaudit.facade.Auditor;
 import com.ericsson.bss.cassandra.ecaudit.facade.DefaultAuditor;
@@ -34,17 +41,22 @@ import com.ericsson.bss.cassandra.ecaudit.filter.role.RoleAuditFilter;
 import com.ericsson.bss.cassandra.ecaudit.filter.yaml.YamlAuditFilter;
 import com.ericsson.bss.cassandra.ecaudit.filter.yamlandrole.YamlAndRoleAuditFilter;
 import com.ericsson.bss.cassandra.ecaudit.logger.AuditLogger;
+import com.ericsson.bss.cassandra.ecaudit.logger.ChronicleAuditLogger;
 import com.ericsson.bss.cassandra.ecaudit.logger.Slf4jAuditLogger;
 import com.ericsson.bss.cassandra.ecaudit.obfuscator.AuditObfuscator;
 import com.ericsson.bss.cassandra.ecaudit.obfuscator.PasswordObfuscator;
 import org.apache.cassandra.auth.IAuthenticator;
 import org.apache.cassandra.config.DatabaseDescriptor;
+import org.apache.cassandra.config.ParameterizedClass;
 import org.apache.cassandra.exceptions.ConfigurationException;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
+@RunWith(MockitoJUnitRunner.StrictStubs.class)
 public class TestAuditAdapterFactory
 {
     @BeforeClass
@@ -75,7 +87,7 @@ public class TestAuditAdapterFactory
 
     @Ignore // Initialization error in Cache
     @Test
-    public void testLoadDefaultWithoutErrorHasExpectedTypes() throws Exception
+    public void testLoadDefaultWithoutErrorHasExpectedType() throws Exception
     {
         AuditAdapter adapter = AuditAdapterFactory.createAuditAdapter();
 
@@ -89,7 +101,7 @@ public class TestAuditAdapterFactory
     }
 
     @Test
-    public void testLoadYamlWithoutErrorHasExpectedTypes() throws Exception
+    public void testLoadYamlFilterWithoutErrorHasExpectedType() throws Exception
     {
         System.setProperty(AuditAdapterFactory.FILTER_TYPE_PROPERTY_NAME, AuditAdapterFactory.FILTER_TYPE_YAML);
 
@@ -99,14 +111,12 @@ public class TestAuditAdapterFactory
         assertThat(auditor).isInstanceOf(DefaultAuditor.class);
 
         DefaultAuditor defaultAuditor = (DefaultAuditor) auditor;
-        assertThat(loggerIn(defaultAuditor)).isInstanceOf(Slf4jAuditLogger.class);
         assertThat(filterIn(defaultAuditor)).isInstanceOf(YamlAuditFilter.class);
-        assertThat(obfuscatorIn(defaultAuditor)).isInstanceOf(PasswordObfuscator.class);
     }
 
     @Ignore // Initialization error in Cache
     @Test
-    public void testLoadRoleWithoutErrorHasExpectedTypes() throws Exception
+    public void testLoadRoleFilterWithoutErrorHasExpectedType() throws Exception
     {
         System.setProperty(AuditAdapterFactory.FILTER_TYPE_PROPERTY_NAME, AuditAdapterFactory.FILTER_TYPE_ROLE);
 
@@ -116,14 +126,12 @@ public class TestAuditAdapterFactory
         assertThat(auditor).isInstanceOf(DefaultAuditor.class);
 
         DefaultAuditor defaultAuditor = (DefaultAuditor) auditor;
-        assertThat(loggerIn(defaultAuditor)).isInstanceOf(Slf4jAuditLogger.class);
         assertThat(filterIn(defaultAuditor)).isInstanceOf(RoleAuditFilter.class);
-        assertThat(obfuscatorIn(defaultAuditor)).isInstanceOf(PasswordObfuscator.class);
     }
 
     @Ignore // Initialization error in Cache
     @Test
-    public void testLoadYamlAndRoleWithoutErrorHasExpectedTypes() throws Exception
+    public void testLoadYamlAndRoleFilterWithoutErrorHasExpectedType() throws Exception
     {
         System.setProperty(AuditAdapterFactory.FILTER_TYPE_PROPERTY_NAME, AuditAdapterFactory.FILTER_TYPE_YAML_AND_ROLE);
 
@@ -133,13 +141,11 @@ public class TestAuditAdapterFactory
         assertThat(auditor).isInstanceOf(DefaultAuditor.class);
 
         DefaultAuditor defaultAuditor = (DefaultAuditor) auditor;
-        assertThat(loggerIn(defaultAuditor)).isInstanceOf(Slf4jAuditLogger.class);
         assertThat(filterIn(defaultAuditor)).isInstanceOf(YamlAndRoleAuditFilter.class);
-        assertThat(obfuscatorIn(defaultAuditor)).isInstanceOf(PasswordObfuscator.class);
     }
 
     @Test
-    public void testLoadNoneWithoutErrorHasExpectedTypes() throws Exception
+    public void testLoadNoFilterWithoutErrorHasExpectedType() throws Exception
     {
         System.setProperty(AuditAdapterFactory.FILTER_TYPE_PROPERTY_NAME, AuditAdapterFactory.FILTER_TYPE_NONE);
 
@@ -149,24 +155,86 @@ public class TestAuditAdapterFactory
         assertThat(auditor).isInstanceOf(DefaultAuditor.class);
 
         DefaultAuditor defaultAuditor = (DefaultAuditor) auditor;
-        assertThat(loggerIn(defaultAuditor)).isInstanceOf(Slf4jAuditLogger.class);
         assertThat(filterIn(defaultAuditor)).isInstanceOf(DefaultAuditFilter.class);
-        assertThat(obfuscatorIn(defaultAuditor)).isInstanceOf(PasswordObfuscator.class);
     }
 
     @Test
-    public void testLoadUnknownFails()
+    public void testLoadUnknownFilterFails()
     {
         System.setProperty(AuditAdapterFactory.FILTER_TYPE_PROPERTY_NAME, "UNKNOWN");
 
         assertThatExceptionOfType(ConfigurationException.class)
-        .isThrownBy(AuditAdapterFactory::createAuditAdapter);
+        .isThrownBy(AuditAdapterFactory::createAuditAdapter)
+        .withMessageContaining("Unrecognized audit filter type")
+        .withMessageContaining("UNKNOWN");
+    }
+
+    @Ignore // Initialization error in Cache
+    @Test
+    public void testLoadSlf4jLogger() throws Exception
+    {
+        AuditConfig auditConfig = givenAuditConfig("com.ericsson.bss.cassandra.ecaudit.logger.Slf4jAuditLogger", Collections.emptyMap());
+
+        AuditAdapter adapter = AuditAdapterFactory.createAuditAdapter(auditConfig);
+
+        Auditor auditor = auditorIn(adapter);
+        assertThat(auditor).isInstanceOf(DefaultAuditor.class);
+
+        DefaultAuditor defaultAuditor = (DefaultAuditor) auditor;
+        assertThat(loggerIn(defaultAuditor)).isInstanceOf(Slf4jAuditLogger.class);
+    }
+
+    @Ignore // Initialization error in Cache
+    @Test
+    public void testLoadChronicleLogger() throws Exception
+    {
+        File tempDir = Files.createTempDir();
+        tempDir.deleteOnExit();
+        AuditConfig auditConfig = givenAuditConfig("com.ericsson.bss.cassandra.ecaudit.logger.ChronicleAuditLogger", ImmutableMap.of("log_dir", tempDir.getPath()));
+
+        AuditAdapter adapter = AuditAdapterFactory.createAuditAdapter(auditConfig);
+
+        Auditor auditor = auditorIn(adapter);
+        assertThat(auditor).isInstanceOf(DefaultAuditor.class);
+
+        DefaultAuditor defaultAuditor = (DefaultAuditor) auditor;
+        assertThat(loggerIn(defaultAuditor)).isInstanceOf(ChronicleAuditLogger.class);
+    }
+
+    @Test
+    public void testLoadChronicleLoggerWithoutMandatoryParameters()
+    {
+        AuditConfig auditConfig = givenAuditConfig("com.ericsson.bss.cassandra.ecaudit.logger.ChronicleAuditLogger", Collections.emptyMap());
+
+        assertThatExceptionOfType(ConfigurationException.class)
+        .isThrownBy(() -> AuditAdapterFactory.createAuditAdapter(auditConfig))
+        .withMessageContaining("Audit logger backend failed at initialization")
+        .withMessageContaining("log_dir");
+    }
+
+    @Test
+    public void testLoadInvalidLogger()
+    {
+        AuditConfig auditConfig = givenAuditConfig("no.working.InvalidAuditLogger", Collections.emptyMap());
+
+        assertThatExceptionOfType(ConfigurationException.class)
+        .isThrownBy(() -> AuditAdapterFactory.createAuditAdapter(auditConfig))
+        .withMessageContaining("Failed to initialize audit logger backend")
+        .withMessageContaining("InvalidAuditLogger");
     }
 
     private static String getPathToTestResourceFile()
     {
         URL url = TestAuditAdapterFactory.class.getResource("/mock_configuration.yaml");
         return url.getPath();
+    }
+
+    private static AuditConfig givenAuditConfig(String s, Map<String, String> parameters)
+    {
+        ParameterizedClass parameterizedClass = new ParameterizedClass(s, parameters);
+        AuditConfig auditConfig = mock(AuditConfig.class);
+        when(auditConfig.getLoggerBackendParameters()).thenReturn(parameterizedClass);
+        return auditConfig;
     }
 
     private static Auditor auditorIn(AuditAdapter adapter) throws Exception